refactor: Change the get_current_user functionality of the file
