+++ backend/app/services/user_services.py
@@ -3,23 +3,37 @@ from app.models.user_model import UserCreate, UserOut, UserUpdate
 from app.database.supabase_client import supabase as supabase_Configured
 from uuid import UUID
 from typing import Optional
-
-from app.services.auth_services import get_current_user
+from fastapi.security import OAuth2PasswordBearer
+from jose import JWTError, jwt
+from app.config import settings
 
 supabase = supabase_Configured
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")
 
 def create_user(user_data: UserCreate) -> UserOut:
   response = supabase.table("users").insert(user_data.model_dump()).execute()
-
   if response.error:
     raise Exception(f"Error creating user: {response.error.message}")
   
   created_user = response.data[0]
   return UserOut(**created_user)
 
+def get_current_user(token: str = Depends(oauth2_scheme)) -> UserOut:
+  try:
+    payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
+    user_id: str = payload.get("sub")
+    if user_id is None:
+      raise HTTPException(status_code=401, detail="Invalid credentials")
+  except JWTError:
+    raise HTTPException(status_code=401, detail="Invalid token")
+
+  user = get_user_by_id(user_id)
+  if user is None:
+    raise HTTPException(status_code=404, detail="User not found")
+  return user
+
 def get_user_by_email(email: str) -> Optional[UserOut]:
-  response = supabase.table("users").select("*").eq("email", email).single().execute()
-  
+  response = supabase.table("users").select("*").eq("email", email).execute()
   if response.error:
     return None
